pyproject.toml
[project]
name = "finanzas-personales-inteligentes"
version = "0.1.0"
description = "WhatsApp-first personal finance assistant focused on Travel Mode with multi-modal expense tracking"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    # Validation & Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    # LangChain Core & Orchestration
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    # LLM Providers (Multi-provider support)
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=0.1.0",
    # LlamaIndex & LlamaParse (Receipt/Document parsing)
    "llama-index-core>=0.11.0",
    "llama-parse>=0.5.0",
    # OpenAI SDK (Whisper API)
    "openai>=1.50.0",
    # Object Storage
    "minio>=7.2.0",
    # Logging
    "structlog>=24.1.0",
    "python-json-logger>=2.0.0",
    # File Processing
    "python-multipart>=0.0.6",
    "pillow>=10.2.0",
    "python-magic>=0.4.27",
    "python-dateutil>=2.8.0",
    "requests>=2.31.0",
    # Optional: Local Whisper (heavier, offline-capable)
    "faster-whisper>=1.0.0",
    # Optional: Langfuse (Future prompt management)
    "langfuse>=2.0.0",
    # Qdrant Vector Database (for Vanna RAG)
    "qdrant-client>=1.12.0",
    # SQL Parsing & Validation
    "sqlparse>=0.5.0",
    "vanna>=2.0",
    # WhatsApp Integration (Twilio)
    "twilio>=9.0.0",
    "phonenumbers>=8.13.0",
    # HTTP Client (for media downloads)
    "httpx>=0.26.0",
]

# Vanna 2.x dependencies (install separately from git)
# pip install 'vanna[anthropic,qdrant] @ git+https://github.com/vanna-ai/vanna.git@v2'

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=term-missing"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
